+++ src/zenflows/vf/economic_event.ex
@@ -77,10 +77,10 @@ schema "vf_economic_event" do
 	belongs_to :resource_conforms_to, ResourceSpecification
 	field :resource_quantity, :map, virtual: true
 	belongs_to :resource_quantity_has_unit, Unit
-	field :resource_quantity_has_numerical_value, :float
+	field :resource_quantity_has_numerical_value, :decimal
 	field :effort_quantity, :map, virtual: true
 	belongs_to :effort_quantity_has_unit, Unit
-	field :effort_quantity_has_numerical_value, :float
+	field :effort_quantity_has_numerical_value, :decimal
 	field :has_beginning, :utc_datetime_usec
 	field :has_end, :utc_datetime_usec
 	field :has_point_in_time, :utc_datetime_usec
